Add functionality check for Add operator, fix bug
Because don't have right dumbpy module, so intend to write a base
version of numc module, then change it to dumbpy module, and then
improve numc module comparing to initial numc module.
So add functionality check for add operator.

Parameter arg is not a tuple by default, is just a numc.Matric type